split off user-defined control policy and param loading to separate file; split torque_controller.cpp to .hpp/.cpp
